Refactored code executor
