+++ WLC/__main__.py
@@ -2,7 +2,7 @@ import argparse
 import logging
 import tkinter as tk
 
-from WLC.code_executor import CodeExecutor, DEFAULT_DOCKER_PORT
+from WLC.code_executor.executor import CodeExecutor, DEFAULT_DOCKER_PORT
 from WLC.code_fixing.codefixer import CodeFixer
 from WLC.image_processing.camera import Camera
 from WLC.image_processing.preprocessor import Preprocessor
